requirements.txt
qrcode
selenium
pyzbar
pillow